+++ CHANGES.rst
@@ -20,7 +20,8 @@ Deprecated
   Patch by Adam Turner.
 * #13637: Deprecate the :py:meth:`!set_application` method
   of :py:class:`~sphinx.parsers.Parser` objects.
-  Sphinx now directly sets the :py:attr:`!config` and :py:attr:`!env` attributes.
+  Patch by Adam Turner.
+* #13644: Deprecate the :py:attr:`!Parser.config` and :py:attr:`!env` attributes.
   Patch by Adam Turner.
 
 Features added